Convert Badge to TS
