--- src/Badge/Badge.d.ts
@@ -1,10 +1,0 @@
-import { PureComponent } from 'react';
-
-export type BadgeProps = {
-  label: any,
-  active?: boolean,
-  className?: string,
-  style?: object
-}
-
-export default class Badge extends PureComponent<BadgeProps> {}
