--- src/Badge/index.js
@@ -1,2 +1,0 @@
-import Badge from './Badge';
-export default Badge;
